separation of interface listeners
