+++ app/src/main/java/com/assessment/westwingcampaign/ui/adapters/CampaignSingleViewAdapter.kt
@@ -27,7 +27,3 @@ class CampaignSingleViewAdapter(var listener: ItemClickListener) : RecyclerView.
         notifyDataSetChanged()
     }
 }
-interface ItemClickListener {
-    fun navigate()
-    fun callNow()
-}
