+++ app/src/main/java/com/assessment/westwingcampaign/ui/adapters/ItemClickListener.kt
@@ -1,0 +1,6 @@
+package com.assessment.westwingcampaign.ui.adapters
+
+interface ItemClickListener {
+    fun navigate()
+    fun callNow()
+}
