+++ app/src/main/java/com/assessment/westwingcampaign/ui/adapters/ItemZoomListener.kt
@@ -1,0 +1,5 @@
+package com.assessment.westwingcampaign.ui.adapters
+
+interface ItemZoomListener {
+    fun navigate(position: Int)
+}
